+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 * Fixed field synchronized in multiple threads
 * Qol double click to open the file in the editor
 * Qol added static resource supports
+* Qol show icons for item/fluid/block configurators
 * Added more APIs for Transform
 * Added dynamic lang Support